+++ posttest/src/main/java/com/kbtg/bootcamp/posttest/Lottery/LotteryRepository.java
@@ -1,6 +1,5 @@
 package com.kbtg.bootcamp.posttest.Lottery;
 
-import com.kbtg.bootcamp.posttest.User.UserTicket;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
