+++ posttest/src/main/java/com/kbtg/bootcamp/posttest/UserTicket/UserTicket.java
@@ -1,4 +1,4 @@
-package com.kbtg.bootcamp.posttest.User;
+package com.kbtg.bootcamp.posttest.UserTicket;
 
 import com.kbtg.bootcamp.posttest.Lottery.Lottery;
 import jakarta.persistence.*;
