+++ posttest/src/main/java/com/kbtg/bootcamp/posttest/UserTicket/UserTicketRepository.java
@@ -1,4 +1,4 @@
-package com.kbtg.bootcamp.posttest.User;
+package com.kbtg.bootcamp.posttest.UserTicket;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
